Intent: modularize shared metric helpers Operators involved: [Coherence|Self-organization|Resonance] Affected invariants: [#9] Key changes: - move coherence, trig cache, sense index utilities into dedicated metrics modules - update imports across engine and tests and drop metrics_utils module - adjust trig cache reuse test to enforce python fallback when needed Expected risks/dissonances: potential cache invalidation regressions mitigated by targeted tests Metrics: (C(t), Si, νf, phase) unchanged; focused pytest modules pass Equivalence map: metrics_utils.compute_coherence -> metrics.common.compute_coherence; metrics_utils.compute_Si -> metrics.sense_index.compute_Si; metrics_utils.get_trig_cache -> metrics.trigonometry.get_trig_cache
